@@ -10,29 +10,29 @@
   },
   "dependencies": {
     "@aws-sdk/client-ec2": "3.332.0",
-    "@rainbow-me/rainbowkit": "1.0.2",
-    "@tailwindcss/typography": "0.5.9",
+    "@rainbow-me/rainbowkit": "^1.2.0",
+    "@tailwindcss/typography": "^0.5.10",
     "@wagmi/chains": "1.1.0",
-    "countdown": "2.6.0",
-    "goober": "2.1.13",
-    "next": "13.4.0",
-    "react": "18.2.0",
+    "countdown": "^2.6.0",
+    "goober": "^2.1.13",
+    "next": "13.4.19",
+    "react": "^18.2.0",
     "react-dom": "18.2.0",
-    "viem": "1.1.3",
-    "wagmi": "1.2.1"
+    "viem": "^1.18.8",
+    "wagmi": "^1.4.5"
   },
   "devDependencies": {
-    "@headlessui/react": "1.7.14",
-    "@heroicons/react": "2.0.18",
-    "@types/node": "17.0.31",
-    "@types/react": "18.0.9",
-    "@types/react-dom": "18.0.3",
-    "@wagmi/cli": "1.0.1",
+    "@headlessui/react": "^1.7.17",
+    "@heroicons/react": "^2.0.18",
+    "@types/node": "^20.8.10",
+    "@types/react": "^18.2.36",
+    "@types/react-dom": "^18.2.14",
+    "@wagmi/cli": "^1.5.2",
     "autoprefixer": "10.4.14",
-    "eslint": "8.15.0",
-    "eslint-config-next": "12.1.6",
-    "postcss": "8.4.24",
-    "tailwindcss": "3.3.2",
-    "typescript": "5.0.4"
+    "eslint": "^8.53.0",
+    "eslint-config-next": "^14.0.1",
+    "postcss": "^8.4.31",
+    "tailwindcss": "^3.3.5",
+    "typescript": "^5.2.2"
   }
 }
