@@ -25,22 +25,22 @@
   },
   "homepage": "https://heroicus.xyz",
   "dependencies": {
-    "@openzeppelin/contracts": "^4.9.0-rc.1",
-    "dotenv": "^16.0.0",
-    "typechain": "^8.1.1",
-    "typescript": "^4.6.4"
+    "@openzeppelin/contracts": "^5.0.0",
+    "dotenv": "^16.3.1",
+    "typechain": "^8.3.2",
+    "typescript": "^5.2.2"
   },
   "devDependencies": {
-    "@nomicfoundation/hardhat-chai-matchers": "^1.0.6",
-    "@nomicfoundation/hardhat-network-helpers": "^1.0.8",
-    "@nomicfoundation/hardhat-toolbox": "^1.0.2",
-    "@types/chai": "^4.3.5",
-    "@types/node": "^17.0.31",
-    "chai": "^4.3.7",
-    "ethers": "^5.7.2",
-    "hardhat": "^2.14.0",
-    "hardhat-deploy": "^0.11.29",
-    "hardhat-deploy-ethers": "0.3.0-beta.13",
+    "@nomicfoundation/hardhat-chai-matchers": "^2.0.2",
+    "@nomicfoundation/hardhat-network-helpers": "^1.0.9",
+    "@nomicfoundation/hardhat-toolbox": "^3.0.0",
+    "@types/chai": "^4.3.9",
+    "@types/node": "^20.8.10",
+    "chai": "^4.3.10",
+    "ethers": "^6.8.1",
+    "hardhat": "^2.19.0",
+    "hardhat-deploy": "^0.11.43",
+    "hardhat-deploy-ethers": "^0.4.1",
     "sol-merger": "^4.4.0",
     "ts-node": "^10.9.1"
   }
