@@ -45,24 +45,14 @@ export function Index() {
                    style={{'objectFit': 'cover', 'objectPosition': '0 0'} as any}/>
             <div className="">
               <img className="mt-10 inline-block" width="100" height="100" src="/images/fantom.svg"/>
-              <img className="mt-10 inline-block ml-20" src="https://d0.awsstatic.com/logos/powered-by-aws-white.png" width="200"
+              <img className="mt-10 inline-block ml-20" src="https://d0.awsstatic.com/logos/powered-by-aws-white.png"
+                   width="200"
                    height="72"/>
             </div>
 
           </div>
         </div>
       </div>
-
-      {/*<div className="my-10 pt-14 pb-10 text-center">*/}
-      {/*  <div className="text-5xl">Anonymous server rentals</div>*/}
-      {/*  <div className="mt-5 text-3xl">Pay with crypto on the Fantom network</div>*/}
-      {/*  <div className="mt-10">*/}
-      {/*    <img className="inline-block" width="100" height="100" src="/images/fantom.svg"/>*/}
-      {/*    <img className="ml-20 inline-block" src="https://d0.awsstatic.com/logos/powered-by-aws-white.png" width="200" height="72"/>*/}
-
-      {/*  </div>*/}
-      {/*</div>*/}
-
       <div className="container mx-auto">
 
 
@@ -72,17 +62,10 @@ export function Index() {
                    style={{height: '400px', 'objectFit': 'cover', 'objectPosition': '50% 50%'} as any}/>
           </div>
           <div className="md:ml-[30%] p-20">
-            {/*<div className="text-lg">use case 1:</div>*/}
             <div className="text-5xl font-bold">Stable Diffusion</div>
             <div className="text-3xl mt-5">GPU Servers to generate AI Images</div>
             <div className="text-xl mt-5">Featuring Telsa T4 16GB GPUs</div>
             <div className="text-xl mt-5">Automatic1111 software preinstalled</div>
-            {/*<div className="text-xl mt-3">Text-to-Image, Image-to-Image, Inpainting, ControlNet, and more</div>*/}
-            {/*<div className="text-xl mt-3">Popular models preinstalled</div>*/}
-            {/*<Link href="/portal"*/}
-            {/*      className="text-4xl inline-block border-amber-50 border-4 px-10 py-3 bg-neutral-900 mt-10 hover:border-amber-200">*/}
-            {/*  Run*/}
-            {/*</Link>*/}
           </div>
         </div>
 
@@ -92,14 +75,9 @@ export function Index() {
                    style={{height: '400px', 'objectFit': 'cover', 'objectPosition': '50% 50%'} as any}/>
           </div>
           <div className="md:ml-[30%] p-20">
-            {/*<div className="text-lg">use case 1:</div>*/}
             <div className="text-5xl font-bold">Docker</div>
             <div className="text-3xl mt-5">Burstable CPU servers to host anything using docker containers</div>
             <div className="text-xl mt-5">Portainer for managing docker images/containers</div>
-            {/*<Link href="/portal"*/}
-            {/*      className="text-4xl inline-block border-amber-50 border-4 px-10 py-3 bg-neutral-900 mt-10 hover:border-amber-200">*/}
-            {/*  Run*/}
-            {/*</Link>*/}
           </div>
 
         </div>
@@ -109,12 +87,13 @@ export function Index() {
             <Image src={discord} height="50" width="50" alt="discord logo" className="inline-block mb-2"/>
             <div>Discord</div>
           </a>
-          <a href="https://github.com" target="_blank" className="mx-10 inline-block">
+          <a href="https://github.com/randombits-dev/heroicus" target="_blank" className="mx-10 inline-block">
             <Image src={github} height="50" width="50" alt="github logo" className="inline-block mb-2"/>
             <div>Github</div>
           </a>
           <div>
-            <div className="mt-10">Brought to you by <a className="underline" href="https://randombits.dev">Random Bits</a></div>
+            <div className="mt-10">Brought to you by <a className="underline" href="https://randombits.dev">Random
+              Bits</a></div>
 
           </div>
         </div>
